+++ scripts/4.1_edit_export.py
@@ -18,6 +18,5 @@ model.save("temp/change_project_name.ifc")
 print("\nExport selected entities")
 print("=" * 53 + "\n")
 
-# TODO: re-enable export functionality
-# window = model.get_entities_by_type("IfcWindow")[0]
-# model.export([window], "temp/selected_entities.ifc")
+window = model.get_entities_by_type("IfcWindow")[0]
+model.export("temp/selected_entities.ifc", [window])