+++ src/compas_ifc/file.py
@@ -40,6 +40,14 @@ class IFCFile(object):
         if load_geometries and filepath is not None:
             self.load_geometries()
 
+    @property
+    def schema(self):
+        return self._schema
+
+    @property
+    def schema_name(self):
+        return self.schema.name()
+
     def ensure_classes_generated(self):
         try:
             from compas_ifc.entities.generated import IFC2X3  # noqa: F401
@@ -162,6 +170,69 @@ class IFCFile(object):
 
     def save(self, path):
         self._file.write(path)
+
+    def export(self, path: str, entities: list[Base] = [], as_snippet: bool = False, export_materials: bool = True, export_properties: bool = True, export_styles: bool = True):
+
+        new_file = IFCFile(None, schema=self.schema_name)
+
+        exported = {}
+
+        def export_entity(entity: Base, file: IFCFile):
+
+            if entity.is_a("IfcOwnerHistory"):
+                return file.default_owner_history
+
+            if entity in exported:
+                return exported[entity]
+
+            parent = getattr(entity, "parent", None)
+            if parent and not as_snippet:
+                new_parent_entity = export_entity(parent, file)
+
+            new_entity = file._create_entity(entity.is_a())
+            exported[entity] = new_entity
+
+            if parent and not as_snippet:
+                if hasattr(entity, "ContainedInStructure"):
+                    file._create_entity("IfcRelContainedInSpatialStructure", RelatingStructure=new_parent_entity, RelatedElements=[new_entity])
+                else:
+                    file._create_entity("IfcRelAggregates", RelatingObject=new_parent_entity, RelatedObjects=[new_entity])
+
+            for key, attr in entity.attributes.items():
+                if isinstance(attr, Base):
+                    attr = export_entity(attr, file)
+                elif isinstance(attr, (list, tuple)) and all(isinstance(a, Base) for a in attr):
+                    attr = [export_entity(a, file) for a in attr]
+                
+                try:
+                    setattr(new_entity, key, attr)
+                except TypeError as e:
+                    if e.args[0] == "Unable to set derived attribute":
+                        pass # TODO: follow up with ifcopenshell why this is the case.
+                    else:
+                        print("Failed setting", entity, key, attr, " Error message:", e)
+
+            if export_properties and hasattr(entity, "properties"):
+                new_entity.properties = entity.properties
+
+            if export_materials and hasattr(entity, "HasAssociations"):
+                # TODO: create material settor on class extension.
+                for relation in entity.HasAssociations():
+                    if relation.is_a("IfcRelAssociatesMaterial"):
+                        new_material_entity = export_entity(relation.RelatingMaterial, file)
+                        file._create_entity("IfcRelAssociatesMaterial", RelatingMaterial=new_material_entity, RelatedObjects=[new_entity])
+
+            if export_styles and hasattr(entity, "StyledByItem"):
+                # TODO: create style settor on class extension.
+                for style_item in entity.StyledByItem():
+                    export_entity(style_item, file)
+
+            return new_entity
+
+        for entity in entities:
+            export_entity(entity, new_file)
+
+        new_file.save(path)
 
     def create(self, cls="IfcBuildingElementProxy", parent=None, geometry=None, frame=None, properties=None, **kwargs):
         if isinstance(cls, type):