+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `export` method to `IFCFile` and `Model` to export selected list of entities.
+
 ### Changed
 
 ### Removed